pyproject.toml
[tool.poetry]
name = "steam-workshop-downloader"
version = "0.1.0"
description = ""
authors = ["Алексей Чистов <aleksey.chistov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"

structlog = "^21.1.0"
colorama = "^0.4.4"

aiohttp = "^3.7.4"
aiofiles = "^0.7.0"

requests = "^2.26.0"
beautifulsoup4 = "^4.9.3"

schema = "^0.7.4"

PyYAML = "^5.4.1"
PyInquirer = "^1.0.3"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
isort = "^5.9.2"
wemake-python-styleguide = "^0.15.3"
rope = "^0.19.0"

mypy = "^0.910"

types-aiofiles = "^0.1.7"
types-requests = "^2.25.0"
types-PyYAML = "^5.4.3"

pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
